s13/assignment/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
timm==1.0.12
torch==2.3.1
torchvision==0.18.1
Pillow==10.4.0
requests==2.32.3
httpx==0.27.2
